Move per-game clock rating calculation to common library
- Create common/src/ratings/clockRating.ts with calculateTimeRating(),
  getPerfectLineSeconds(), ClockDatum interface, and constants
- Replace frontend clockRating.ts with re-exports from common
- Remove unused _side parameter from calculateTimeRating calls
